Handle ReadTimeouts when a machine is down
The SF API times out if a machine is down when requesting job
information. As the call is wrapped in tenacity it can take
some time for the retries to fail. This can slow the whole
monitoring loop. To prevent this we now check the machine status
before requesting the jobs.